Move the test to PresearcherTestBase and remove checks for QueriesRunCount
